move websocket logic
